Renamed as 23andmeparser.py
Originally, geneFinder and genePageScraper each had a bug that cancelled the other out. I initially fixed the bug in geneFinder, but did not fix the bug in genePageScraper, so my code broke. I realized that position is an awful metric to use (23andme does not automatically sort by position) so I removed it and fixed the if else in lines 89-92 of the code. I have re-uploaded the correct file as 23andmeparser. I have made a few other edits to the other functions to protect against possible errors.